moved to integer rating

+++ src/main/java/com/iota/iri/service/viewModels/TransactionViewModel.java
@@ -237,7 +237,7 @@ public class TransactionViewModel {
         update("rating");
     }
 
-    public long getRating() {
+    public int getRating() {
         return transaction.rating;
     }
